add filter, Returnbill,hide return and delete button and fininsh it 99%

+++ src/component/sales/invoice/index.js
@@ -4,11 +4,18 @@ import { http } from '../../../axios'
 import { MyDate } from '../../../Custom/MyDate'
 import './index.css'
 import PrintButton from './PrintButton'
-import {SpeInvoiceNumber} from '../../../Custom/SpeInvoiceNumber'
+import { SpeInvoiceNumber } from '../../../Custom/SpeInvoiceNumber'
+import Filter from '../../../Custom/Filter'
 
 function Invoice() {
     const [sales, setSales] = useState([])
-    const [reload,setReload] = useState()
+    const [reload, setReload] = useState()
+
+    const [filterData, setfilterData] = useState()
+
+    const FilterData = (filter) => {
+        setfilterData(filter)
+    }
 
     // const ComponentRef = useRef()
 
@@ -25,25 +32,42 @@ function Invoice() {
     let total = 0, gst = 0, nop = 0, count = 0, x
 
     for (x of sales) {
-        total += x.total
-        gst += x.gst
-        nop += x.numofproduct
-        count += 1
+        if (x.success) {
+            total += x.total
+            gst += x.gst
+            nop += x.numofproduct
+            count += 1
+        }
     }
 
     const CancelBill = (id) => {
-        console.log("deletebill", id)
-        http.delete("/sales", { params: { id: id } })
-            .then(res => {
-               alert(res.data)
-               setReload(id)
-            })
-            .catch(err => {
-                console.log("Error", err)
-            })
+        if (window.confirm("Confirm to Delete")) {
+            http.delete("/sales", { params: { id: id } })
+                .then(res => {
+                    alert(res.data)
+                    setReload(id)
+                })
+                .catch(err => {
+                    console.log("Error", err)
+                })
+        }
 
     }
-    
+
+    const ReturnBill = (id) => {
+        console.log(id)
+        if (window.confirm("Confirm to Return")) {
+            http.put("/sales", { id: id })
+                .then(res => {
+                    alert(res.data)
+                    setReload(id)
+                })
+                .catch(err => {
+                    console.log("Error", err)
+                })
+        }
+    }
+
 
     return (
         <div>
@@ -69,78 +93,170 @@ function Invoice() {
 
             </div>
 
-            <div className="w3-container">
+
+            <div className="sFilter w3-container mt-3">
                 {
-                    sales.map((data, index) => (
-                        <div key={index}>
-                            <div className="w3-card-2 m-3 scard">
-
-                                <div className="scardtop">
-
-                                    <div className="saddress">
-                                        <div className="sbilladd">
-                                            <p><b>Bill To </b></p>
-                                            <p>{data.bill.name}</p>
-                                        </div>
-                                        <div className="sbilladd">
-                                            <p><b>Ship To </b></p>
-                                            <p>{data.ship.name}</p>
-                                        </div>                                        
-                                    </div>
-
-                                    <div className="billaction">
-                                        <div className="w3-center sbuttons">
-                                            <div className="sview">
-                                                <button className="w3-button w3-blue mt-3" data-toggle="modal" data-target={`#id${data._id}`}>View</button>
-                                            </div>
-                                            <div className="sdelete">
-                                                <button className="w3-button w3-red mt-3" onClick={()=>CancelBill(data._id)} >Delete</button>
-                                            </div>
-                                        </div>
-                                    </div>
-
-                                </div>
-
-                                <div className="scarddown">
-                                    <div>
-                                        <p className="pl-2">Invoice No : {SpeInvoiceNumber(data.invoicenumber)}</p>
-                                    </div>
-                                    <div>
-                                        <p>Date : {MyDate(data.invoicedate)}</p>
-                                    </div>
-                                    <div>
-                                        <p>Payment : {data.payment}</p>
-                                    </div>
-                                    <div>
-                                        <p>Gst : {data.gst.toFixed(2)}</p>
-                                    </div>
-                                    <div>
-                                        <p>Total : {data.total.toFixed(2)}</p>
-                                    </div>
-                                    <div>
-                                        <p className="pr-2">NOP : {data.numofproduct}</p>
-                                    </div>
-                                </div>
-                            </div>
-
-
-
-
-                            <div className="modal" id={`id${data._id}`} >
-                                <div className="modal-dialog modal-xl">
-                                    <div className="modal-content">
-                                        {/* <SalesBill SalesData={data} ref={ComponentRef} />
-                                        <br></br> */}
-                                        <PrintButton SalesData={data} />                                        
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-
-
-                    ))
+                    <Filter FilterData={FilterData} data={sales} />
                 }
+            </div>
 
+            <div className="w3-container">
+
+                {
+                    filterData ? (
+                        filterData.map((data, index) =>
+                            <div key={index}>
+                                <div className="w3-card-2 m-3 scard">
+
+                                    <div className="scardtop">
+
+                                        <div className="saddress">
+                                            <div className="sbilladd">
+                                                <p><b>Bill To </b></p>
+                                                <p>{data.bill.name}</p>
+                                            </div>
+                                            <div className="sbilladd">
+                                                <p><b>Ship To </b></p>
+                                                <p>{data.ship.name}</p>
+                                            </div>
+                                        </div>
+
+                                        <div className="billaction">
+                                            <div className="w3-center sbuttons">
+                                                <div className="sview">
+                                                    <button className="w3-button w3-blue mt-3" data-toggle="modal" data-target={`#id${data._id}`}>View</button>
+                                                </div>
+                                                {
+                                                    data.success && (
+                                                        <>
+                                                            <div className="sdelete">
+                                                                <button className="w3-button w3-yellow mt-3" onClick={() => ReturnBill(data._id)} >Return</button>
+                                                            </div>
+                                                            <div className="sdelete">
+                                                                <button className="w3-button w3-red mt-3" onClick={() => CancelBill(data._id)} >Delete</button>
+                                                            </div>
+                                                        </>
+                                                    )
+                                                }
+
+                                            </div>
+                                        </div>
+
+                                    </div>
+
+                                    <div className="scarddown">
+                                        <div>
+                                            <p className="pl-2">Invoice No : {SpeInvoiceNumber(data.invoicenumber)}</p>
+                                        </div>
+                                        <div>
+                                            <p>Date : {MyDate(data.invoicedate)}</p>
+                                        </div>
+                                        <div>
+                                            <p>Payment : {data.payment}</p>
+                                        </div>
+                                        <div>
+                                            <p>Gst : {data.gst.toFixed(2)}</p>
+                                        </div>
+                                        <div>
+                                            <p>Total : {data.total.toFixed(2)}</p>
+                                        </div>
+                                        <div>
+                                            <p className="pr-2">NOP : {data.numofproduct}</p>
+                                        </div>
+                                    </div>
+                                </div>
+
+                                <div className="modal" id={`id${data._id}`} >
+                                    <div className="modal-dialog modal-xl">
+                                        <div className="modal-content">
+                                            <PrintButton SalesData={data} />
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+
+
+                        )
+
+                    )
+                        :
+                        (
+                            sales.map((data, index) => (
+                                <div key={index}>
+                                    <div className="w3-card-2 m-3 scard">
+
+                                        <div className="scardtop">
+
+                                            <div className="saddress">
+                                                <div className="sbilladd">
+                                                    <p><b>Bill To </b></p>
+                                                    <p>{data.bill.name}</p>
+                                                </div>
+                                                <div className="sbilladd">
+                                                    <p><b>Ship To </b></p>
+                                                    <p>{data.ship.name}</p>
+                                                </div>
+                                            </div>
+
+                                            <div className="billaction">
+                                                <div className="w3-center sbuttons">
+                                                    <div className="sview">
+                                                        <button className="w3-button w3-blue mt-3" data-toggle="modal" data-target={`#id${data._id}`}>View</button>
+                                                    </div>
+                                                    {
+                                                        data.success && (
+                                                            <>
+                                                                <div className="sdelete">
+                                                                    <button className="w3-button w3-yellow mt-3" onClick={() => ReturnBill(data._id)} >Return</button>
+                                                                </div>
+                                                                <div className="sdelete">
+                                                                    <button className="w3-button w3-red mt-3" onClick={() => CancelBill(data._id)} >Delete</button>
+                                                                </div>
+                                                            </>
+                                                        )
+                                                    }
+
+                                                </div>
+                                            </div>
+
+                                        </div>
+
+                                        <div className="scarddown">
+                                            <div>
+                                                <p className="pl-2">Invoice No : {SpeInvoiceNumber(data.invoicenumber)}</p>
+                                            </div>
+                                            <div>
+                                                <p>Date : {MyDate(data.invoicedate)}</p>
+                                            </div>
+                                            <div>
+                                                <p>Payment : {data.payment}</p>
+                                            </div>
+                                            <div>
+                                                <p>Gst : {data.gst.toFixed(2)}</p>
+                                            </div>
+                                            <div>
+                                                <p>Total : {data.total.toFixed(2)}</p>
+                                            </div>
+                                            <div>
+                                                <p className="pr-2">NOP : {data.numofproduct}</p>
+                                            </div>
+                                        </div>
+                                    </div>
+
+                                    <div className="modal" id={`id${data._id}`} >
+                                        <div className="modal-dialog modal-xl">
+                                            <div className="modal-content">
+                                                <PrintButton SalesData={data} />
+                                            </div>
+                                        </div>
+                                    </div>
+                                </div>
+
+
+                            )
+                            )
+                        )
+                }
             </div>
         </div>
     )
